Python redo of PVMonitor project to read the output from an Outback Power Systems Photovoltaic system
This the inital commit of this new project.  It is currently requiring python 2.7.   The data from the PV system is sent from an embedded server.  An example of the output of the server is found in the matetest.obm file that contains the raw data from one day's output.  Currently the program reads the stream from the server.  A test bed that simulates the server and streams the data from the .obm file is in progress.

The program is not complete.  Some of the data is currently being decoded.  The partially decoded output is blocks of data that are received once per second:

06172019123857
{'Op Mode': 'Sell Enabled', 'Sell Current': '17', 'Error Mode': 'OK', 'ACIn Volt': '125', 'Battery Volt': 52.0, 'Inverter Current': '20', 'Warn Mode': '128', 'Buy Current': '00', 'ID': '1', 'ACOut Volt': '125', 'Chgrger Current': '00', 'Misc': '024', 'AC Mode': '02'}
{'Op Mode': 'Sell Enabled', 'Sell Current': '07', 'Error Mode': 'OK', 'ACIn Volt': '126', 'Battery Volt': 52.0, 'Inverter Current': '16', 'Warn Mode': '000', 'Buy Current': '00', 'ID': '2', 'ACOut Volt': '124', 'Chgrger Current': '00', 'Misc': '152', 'AC Mode': '02'}
{'Op Mode': 'Sell Enabled', 'Sell Current': '15', 'Error Mode': 'OK', 'ACIn Volt': '126', 'Battery Volt': 52.0, 'Inverter Current': '15', 'Warn Mode': '000', 'Buy Current': '00', 'ID': '3', 'ACOut Volt': '125', 'Chgrger Current': '00', 'Misc': '024', 'AC Mode': '02'}
{'Battery Volt': '519', 'charger Current': '38', 'Power (KWH)': '044', 'Aux Mode': '00', 'Error': '000', 'Charger Mode': '01', 'ID': 'E', 'PV Voltage': '060', 'PV Current': '35'}
{'Battery Volt': '520', 'charger Current': '36', 'Power (KWH)': '039', 'Aux Mode': '00', 'Error': '000', 'Charger Mode': '01', 'ID': 'F', 'PV Voltage': '057', 'PV Current': '35'}
{'Battery Volt': '523', 'charger Current': '38', 'Power (KWH)': '044', 'Aux Mode': '00', 'Error': '000', 'Charger Mode': '01', 'ID': 'G', 'PV Voltage': '061', 'PV Current': '35'}
{'Battery Volt': '520', 'charger Current': '41', 'Power (KWH)': '051', 'Aux Mode': '00', 'Error': '000', 'Charger Mode': '02', 'ID': 'H', 'PV Voltage': '058', 'PV Current': '39'}

The first line is a timestamp (month/date/year/hour/min/sec
Each line represents  the decoded output of one device on the system.  Inverter data is represented by "ID:1-3" (next 3 lines).  Charge Controller data is represented  by "ID:E-H" (next 4 lines).